+++ lib/getFatcaData-SAMPLE.php
@@ -1,0 +1,7 @@
+<?php
+function getFatcaData() {
+	return array(
+		array("ENT_COD"=>"1234","ENT_FIRSTNAME"=>"Clyde","ENT_LASTNAME"=>"Barrow","ENT_FATCA_ID"=>"123-1234-123","ENT_ADDRESS"=>"Some street somewhere","ResidenceCountry"=>"US","accountsTotalUsd"=>100000000),
+		array("ENT_COD"=>"5678","ENT_FIRSTNAME"=>"Bonnie","ENT_LASTNAME"=>"Parker","ENT_FATCA_ID"=>"456-1234-123","ENT_ADDRESS"=>"Dallas, Texas","ResidenceCountry"=>"US","accountsTotalUsd"=>100)
+	);
+};